vim-patch:8.1.0228, 8.1.1384, 8.1.1386, 8.1.1393, 8.2.3040

+++ src/nvim/fileio.c
@@ -100,8 +100,8 @@ struct bw_info {
   char_u bw_rest[CONV_RESTLEN];  // not converted bytes
   int bw_restlen;                // nr of bytes in bw_rest[]
   int bw_first;                  // first write call
-  char_u *bw_conv_buf;      // buffer for writing converted chars
-  int bw_conv_buflen;            // size of bw_conv_buf
+  char_u *bw_conv_buf;           // buffer for writing converted chars
+  size_t bw_conv_buflen;         // size of bw_conv_buf
   int bw_conv_error;             // set for conversion error
   linenr_T bw_conv_error_lnum;   // first line with error or zero
   linenr_T bw_start_lnum;        // line number at start of buffer
